rename `generate_matrix` module with `decode_matrix`
